Move the new app to a separate folder and entrypoint
