+++ saleor/dashboard/category/urls.py
@@ -3,7 +3,7 @@ from django.conf.urls import url
 from . import views
 
 urlpatterns = [
-    url(r'^',
+    url(r'^$',
         views.category_list, name='category-list'),
     url(r'^(?P<pk>[0-9]+)/$',
         views.category_detail, name='category-detail'),
