+++ saleor/static/dashboard-next/js/category/details/categoryList.js
@@ -6,7 +6,7 @@ import { withRouter } from 'react-router-dom';
 
 import { ListCard } from '../../components/cards';
 import { categoryChildren, rootCategoryChildren } from '../queries';
-import { createQueryString } from '../../../utils';
+import { createQueryString } from '../../misc';
 
 const tableHeaders = [
   {
