+++ saleor/static/dashboard-next/js/category/details/details.js
