+++ saleor/static/dashboard-next/js/category/form/categoryUpdateForm.js
