+++ saleor/static/dashboard-next/js/category/index.js
