+++ saleor/static/dashboard-next/js/components/table.js
