+++ templates/dashboard/next.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+{% load staticfiles i18n %}
+{% load render_bundle from webpack_loader %}
+{% load version %}
+{% load serialize_messages from utils %}
+
+<html lang="{{ LANGUAGE_CODE }}" class="no-js">
+  <head>
+    <title>
+      {% block title %}
+        {% trans "Dashboard - Saleor" context "Dashboard default page title" %}
+      {% endblock %}
+    </title>
+    {% block meta %}
+      <meta charset="utf-8">
+      <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
+      <meta name="theme-color" content="#21915a" />
+      <meta name="robots" content="noindex, nofollow">
+    {% endblock meta %}
+
+    {% include 'favicon.html' %}
+  </head>
+
+  <body>
+    <div id="dashboard-app"></div>
+    <script type="text/javascript" src="{% url 'javascript-catalog' %}"></script>
+    {% render_bundle 'dashboard-next' 'js' %}
+  </body>
+</html>
